Rename pkgsrc to pkglib
