+++ src/cmd/linuxkit/pkglib/git.go
@@ -1,4 +1,4 @@
-package pkgsrc
+package pkglib
 
 // Thin wrappers around git CLI invocations
